backend/requirements.txt
fastapi
uvicorn[standard]
ollama
surrealdb
python-dotenv
pydantic

